+++ dicom_parser/utils/code_strings/scanning_sequence.py
@@ -4,7 +4,7 @@ from dicom_parser.utils.choice_enum import ChoiceEnum
 class ScanningSequence(ChoiceEnum):
     """
     Represents the `Scanning Sequence <https://dicom.innolitics.com/ciods/mr-image/mr-image/00180020>`_ attribute.
-    
+
     """
 
     SE = "Spin Echo"