+++ dicom_parser/utils/sequence_detector/exceptions.py
@@ -1,4 +1,0 @@
-WRONG_DEFINITION_TYPE = (
-    "Sequence definition must be a dict, a list, or a tuple, not {definition_type}!"
-)
-