+++ dicom_parser/utils/sequence_detector/messages.py
@@ -1,0 +1,10 @@
+"""
+Messages to be used by the
+:class:`~dicom_parser.utils.sequence_detector.sequence_detector.SequenceDetector`
+class.
+
+"""
+
+WRONG_DEFINITION_TYPE = (
+    "Sequence definition must be a dict, a list, or a tuple, not {definition_type}!"
+)