+++ dicom_parser/utils/sequence_detector/sequences.py
@@ -1,24 +1,64 @@
+"""
+Known sequences defined by the expected (parsed) data element values from the header.
+
+"""
+
 MR_SEQUENCES = {
-    "localizer": {"ScanningSequence": ["GR"], "SequenceVariant": ["SP", "OSP"]},
-    "mprage": {"ScanningSequence": ["GR", "IR"], "SequenceVariant": ["SK", "SP", "MP"]},
-    "spgr": {"ScanningSequence": ["GR"], "SequenceVariant": ["SK", "SP", "SS"]},
-    "fspgr": {"ScanningSequence": ["GR"], "SequenceVariant": ["SK", "SS"]},
-    "flair": {"ScanningSequence": ["SE", "IR"], "SequenceVariant": ["SK", "SP", "MP"]},
-    "dti": [
-        {"ScanningSequence": ["EP", "SE"], "SequenceVariant": ["NONE"]},
-        {"ScanningSequence": ["EP", "RM"], "SequenceVariant": ["NONE"]},
-    ],
-    "fmri": [
-        {"ScanningSequence": ["EP"], "SequenceVariant": ["SK", "SS"]},
-        {"ScanningSequence": ["EP", "GR"], "SequenceVariant": ["SS"]},
-    ],
-    "irepi": {
-        "ScanningSequence": ["EP", "IR"],
-        "SequenceVariant": ["SK", "SP", "MP", "OSP"],
+    "Localizer": {
+        "ScanningSequence": {"Gradient Recalled"},
+        "SequenceVariant": {"Spoiled", "Oversampling Phase"},
     },
-    "ep2d": {"ScanningSequence": ["EP"], "SequenceVariant": ["SK", "SP"]},
-    "fse": {"ScanningSequence": ["SE"], "SequenceVariant": ["SK"]},
+    "MPRAGE": {
+        "ScanningSequence": {"Gradient Recalled", "Inversion Recovery"},
+        "SequenceVariant": {"Segmented k-Space", "Spoiled", "MAG Prepared"},
+    },
+    "SPGR": {
+        "ScanningSequence": {"Gradient Recalled"},
+        "SequenceVariant": {"Segmented k-Space", "Spoiled", "Steady State"},
+    },
+    "FSPGR": {
+        "ScanningSequence": {"Gradient Recalled"},
+        "SequenceVariant": {"Segmented k-Space", "Steady State"},
+    },
+    "FLAIR": {
+        "ScanningSequence": {"Spin Echo", "Inversion Recovery"},
+        "SequenceVariant": {"Segmented k-Space", "Spoiled", "MAG Prepared"},
+    },
+    "DTI": (
+        {"ScanningSequence": {"Echo Planar", "Spin Echo"}, "SequenceVariant": {"None"}},
+        {
+            "ScanningSequence": {"Echo Planar", "Research Mode"},
+            "SequenceVariant": {"None"},
+        },
+    ),
+    "fMRI": (
+        {
+            "ScanningSequence": {"Echo Planar"},
+            "SequenceVariant": {"Segmented k-Space", "Steady State"},
+        },
+        {
+            "ScanningSequence": {"Echo Planar", "Gradient Recalled"},
+            "SequenceVariant": {"Steady State"},
+        },
+    ),
+    "IR-EPI": {
+        "ScanningSequence": {"Echo Planar", "Inversion Recovery"},
+        "SequenceVariant": {
+            "Segmented k-Space",
+            "Spoiled",
+            "MAG Prepared",
+            "Oversampling Phase",
+        },
+    },
+    "ep2d": {
+        "ScanningSequence": {"Echo Planar"},
+        "SequenceVariant": {"Segmented k-Space", "Spoiled"},
+    },
+    "FSE": {
+        "ScanningSequence": {"Spin Echo"},
+        "SequenceVariant": {"Segmented k-Space"},
+    },
 }
 
 
-SEQUENCES = {"mr": MR_SEQUENCES}
+SEQUENCES = {"Magnetic Resonance": MR_SEQUENCES}